+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/dto/TournamentGetDTO.java
@@ -9,8 +9,8 @@ public class TournamentGetDTO {
     private String tournamentName;
     private String tournamentCode;
     private ParticipantGetDTO winner;
-    private float breakDuration;
-    private float gameDuration;
+    private int breakDuration;
+    private int gameDuration;
     private String startTime;
     private int numberTables;
     private int amountOfPlayers;
@@ -29,10 +29,10 @@ public class TournamentGetDTO {
     public String getTournamentCode() {
         return tournamentCode;
     }
-    public float getBreakDuration() {
+    public int getBreakDuration() {
         return breakDuration;
     }
-    public float getGameDuration() {
+    public int getGameDuration() {
         return gameDuration;
     }
     public String getStartTime() {
@@ -66,10 +66,10 @@ public class TournamentGetDTO {
     public void setTournamentCode(String tournamentCode) {
         this.tournamentCode = tournamentCode;
     }
-    public void setBreakDuration(float breakDuration) {
+    public void setBreakDuration(int breakDuration) {
         this.breakDuration = breakDuration;
     }
-    public void setGameDuration(float gameDuration) {
+    public void setGameDuration(int gameDuration) {
         this.gameDuration = gameDuration;
     }
     public void setStartTime(String startTime) {
